backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
sentence-transformers
faiss-cpu
google-genai
numpy
pydantic
pytest
python-dotenv
